scripts: add Prefetch.hs
To run:

    nix-shell --pure --run "runhaskell scripts/Prefetch.hs"

Prefetch.hs is an experiment of calling directly into the Dhall
library to read spacchetti. Dhall has no concept of a map of values
with uniform type, so we match into the syntax tree manually and then
check each package.

It will use libnix to prefetch the git repositories and output a json
of hashes. The prefetching clones five repositories at a time and
retries on network errors.